Pass test for discounted total revenue in invoices show page

+++ db/migrate/20230113030731_create_bulk_discounts.rb
@@ -1,7 +1,7 @@
 class CreateBulkDiscounts < ActiveRecord::Migration[5.2]
   def change
     create_table :bulk_discounts do |t|
-      t.integer :percent_discount
+      t.float :percent_discount
       t.integer :quantity_threshold
     end
   end
